rename Settings component file to SecondsSettings
