+++ src/components/SecondsSettings.js
